Improve prometheus metrics generation reliability
That means it shouldn't crash with multiple request types to the same
URI anymore.

Additional changes:
 * Add esptherm_version header containing the current commit hash
 * Add constexpr uint to str utility
 * Add esptherm_build_info metric to metrics page
   * This metrics has the labels esptherm_commit, mcu_type,
arduino_version, sdk_version, and cpp_std_version.

+++ src/webhandler.cpp
@@ -126,7 +126,7 @@ bool web::csvHeaderContains(const char *header, const char *value) {
 				&& *cpos != ';') {
 			start = cpos;
 		} else if (start != NULL && (*cpos == ',' || *cpos == ';')) {
-			if (cpos - start == val_len
+			if ((size_t) (cpos - start) == val_len
 					&& strncmp(start, value, val_len) == 0) {
 				return true;
 			}
@@ -137,7 +137,7 @@ bool web::csvHeaderContains(const char *header, const char *value) {
 		}
 	}
 
-	if (start != NULL && cpos - start == val_len
+	if (start != NULL && (size_t) (cpos - start) == val_len
 			&& strncmp(start, value, val_len) == 0) {
 		return true;
 	} else {
